+++ lib/cinegraph_web/live/movie_live/index_v2_components/filters.ex
@@ -11,7 +11,7 @@ defmodule CinegraphWeb.MovieLive.IndexV2Components.Filters do
   alias CinegraphWeb.NeutralV2Components
 
   @primary_sort_keys ~w(release_date score popularity)
-  @basic_filter_keys ~w(search genres decade lists festivals people people_ids rating_preset show_unreleased)
+  @basic_filter_keys ~w(search genres decade lists festivals people rating_preset show_unreleased)
 
   attr :search_term, :string, default: ""
   attr :sort_options, :list, required: true
@@ -247,6 +247,7 @@ defmodule CinegraphWeb.MovieLive.IndexV2Components.Filters do
 
   def active_filter_count(params, scope) when is_map(params) do
     hidden_keys = hidden_filter_keys(scope)
+    params = normalize_people_filter(params)
 
     filter_count =
       @basic_filter_keys
@@ -258,6 +259,18 @@ defmodule CinegraphWeb.MovieLive.IndexV2Components.Filters do
   end
 
   def active_filter_count(_, _), do: 0
+
+  defp normalize_people_filter(params) do
+    cond do
+      filter_value_present?(params["people_ids"]) ->
+        params
+        |> Map.put("people", params["people_ids"])
+        |> Map.delete("people_ids")
+
+      true ->
+        Map.delete(params, "people_ids")
+    end
+  end
 
   def list_param(params, key) when is_map(params) do
     LiveViewHelpers.parse_array_param(params[key])
